Update usecases.html with new binary content
This commit replaces the existing usecases.html file with a new version that contains updated binary content. As a result, the file's index has changed, reflecting the differences in the binary data. The changes may include enhancements or modifications pertinent to the functionality of the application, although specific text changes are not applicable in this case due to the binary nature of the file.