Remove outdated age consistency summary HTML file
This commit deletes the `age_consistency_summary.html` file from the documentation directory. The content of this file included a summary of inconsistencies in age and birthdate data, displaying various metrics and a table of cases detected. Given the potential changes in reporting requirements and the evolution of the project, it has been decided to remove this outdated file to streamline the documentation and avoid confusion. Future reporting or analysis will use updated formats or tools.