Remove patient segmentation summary HTML file
This commit removes the `patient_segmentation_summary.html` file from the documentation. The file provided a summary of patient segmentation information, including statistics on patient demographics and frequencies of visits. As part of a restructuring of the documentation, this summary is no longer required. Future documentation will focus on updated formats and methods of presenting patient segmentation data. Consequently, the removal of this file will help streamline the documentation and eliminate outdated content.